@@ -1,49 +1,62 @@
-// @dart=2.9
 import 'package:flutter/material.dart';
 import 'package:backdrop/backdrop.dart';
 import 'package:flutter_app_than_so_hoc_2/Pages/hangngay/hangngay.dart';
-
+import 'package:flutter_app_than_so_hoc_2/provider/admob/admob_service.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
+import 'package:google_mobile_ads/google_mobile_ads.dart';
 import 'package:intl/intl.dart';
-
-
-// ignore: import_of_legacy_library_into_null_safe
 import 'package:flutter_easyloading/flutter_easyloading.dart';
-
-
 import 'Pages/home/home.dart';
 import 'Pages/introl/intro.dart';
 import 'Pages/setting/setting.dart';
 import 'generated/l10n.dart';
 
-void main() => runApp(new MyApp());
+void main() {
+  WidgetsFlutterBinding.ensureInitialized();
+  MobileAds.instance.initialize();
+  runApp(MyApp());
+}
 
 class MyApp extends StatefulWidget {
   @override
   State<StatefulWidget> createState() {
     // TODO: implement createState
     return _MyAppState();
-    throw UnimplementedError();
   }
 }
 
 class _MyAppState extends State<MyApp> {
   // ignore: non_constant_identifier_names
-  String title_app;
+  String title_app = '';
 
   // ignore: non_constant_identifier_names
-  String title_app_2;
+  String title_app_2 = '';
 
-  String menu_1;
+  String menu_1 = '';
 
-  String menu_2;
+  String menu_2 = '';
 
-  String menu_3;
+  String menu_3 = '';
+
+  BannerAd? _banner;
+  InterstitialAd? _interstitialAd;
 
   @override
   void initState() {
     // TODO: implement initState
     super.initState();
+    _createBannerAd();
+  }
+
+  _createBannerAd() {
+    _banner = BannerAd(
+        size: AdSize.banner,
+        // adUnitId: 'ca-app-pub-5726417511192387/1590401387',
+        adUnitId: 'ca-app-pub-3940256099942544/6300978111',
+        listener: AdMobService.instance.bannerAdListener,
+        request: AdRequest(),
+    );
+    _banner?.load();
   }
 
   int _currentIndex = 0;
@@ -75,7 +88,7 @@ class _MyAppState extends State<MyApp> {
       supportedLocales: S.delegate.supportedLocales,
       initialRoute: '/',
       builder: EasyLoading.init(),
-      home: new BackdropScaffold(
+      home: BackdropScaffold(
         backLayerBackgroundColor: Color(0xffc9a70e),
         frontLayerBorderRadius: const BorderRadius.only(
             topLeft: Radius.circular(50), topRight: Radius.circular(50)),
@@ -110,6 +123,12 @@ class _MyAppState extends State<MyApp> {
           ],
           onTap: (int position) => {setState(() => _currentIndex = position)},
         ),
+        bottomNavigationBar: _banner == null ? SizedBox.shrink() : Container(
+          // margin: const EdgeInsets.only(bottom: 12),
+          height: _banner?.size.height.toDouble(),
+          width: _banner?.size.width.toDouble(),
+          child: AdWidget(ad: _banner!),
+          ),
       ),
     );
   }
