@@ -7,7 +7,8 @@
 // ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
-// ignore_for_file:unused_import, file_names
+// ignore_for_file:unused_import, file_names, avoid_escaping_inner_quotes
+// ignore_for_file:unnecessary_string_interpolations, unnecessary_string_escapes
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -20,33 +21,48 @@ class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'fr';
 
   final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "bieu_do_ngay_sinh" : MessageLookupByLibrary.simpleMessage("Tableau des dates de naissance"),
-    "chon_ngon_ngu" : MessageLookupByLibrary.simpleMessage("Sélectionnez la langue"),
-    "con_so_chu_dao_cua_ban_la" : MessageLookupByLibrary.simpleMessage("Votre numéro fatidique"),
-    "de_xuat" : MessageLookupByLibrary.simpleMessage("Développement proposé"),
-    "giai_thich" : MessageLookupByLibrary.simpleMessage("Expliquez"),
-    "gioi_thieu" : MessageLookupByLibrary.simpleMessage("Introl"),
-    "hay_chon_ngay_sinh_cua_ban" : MessageLookupByLibrary.simpleMessage("Veuillez sélectionner votre date de naissance"),
-    "home" : MessageLookupByLibrary.simpleMessage("Maison"),
-    "introl_1" : MessageLookupByLibrary.simpleMessage("La numérologie est basée sur votre nom et votre date de naissance pour vous prédire de manière complète. Avec la numérologie, voyez comment les nombres importants affectent votre vie."),
-    "introl_2" : MessageLookupByLibrary.simpleMessage("L\'application analysera votre nom et votre date de naissance pour afficher les résultats en un clin d\'œil! En ​​savoir plus sur les numéros de naissance ou les lignes de vie (exprimez votre personnalité et vos perspectives dans la vie). Voyons ce que ces chiffres disent sur toi"),
-    "khuyet_diem" : MessageLookupByLibrary.simpleMessage("Faiblesse"),
-    "moc_cuoc_doi" : MessageLookupByLibrary.simpleMessage("Jalons de la vie"),
-    "muc_dich_cuoc_song" : MessageLookupByLibrary.simpleMessage("Objectifs de vie"),
-    "nghe_nghiep_phu_hop" : MessageLookupByLibrary.simpleMessage("Occupation convenable"),
-    "ngon_ngu" : MessageLookupByLibrary.simpleMessage("Langue"),
-    "noi_bat" : MessageLookupByLibrary.simpleMessage("En vedette"),
-    "phat_trien_theo_so" : MessageLookupByLibrary.simpleMessage("Croissance"),
-    "so" : MessageLookupByLibrary.simpleMessage("Number"),
-    "than_so_hoc_la_gi" : MessageLookupByLibrary.simpleMessage("QU\'EST-CE QUE LA Numérologie?"),
-    "thiet_lap_ngay_sinh" : MessageLookupByLibrary.simpleMessage("set date of birth"),
-    "tong_quan" : MessageLookupByLibrary.simpleMessage("Présentation"),
-    "tu_vi_hom_nay" : MessageLookupByLibrary.simpleMessage("Horoscope today"),
-    "tuoi" : MessageLookupByLibrary.simpleMessage("Ans"),
-    "uu_diem" : MessageLookupByLibrary.simpleMessage("Avantages"),
-    "xem" : MessageLookupByLibrary.simpleMessage("Soumettre"),
-    "xem_ngay_sinh" : MessageLookupByLibrary.simpleMessage("Accueil"),
-    "y_nghia_cac_con_so_tren_bieu_do" : MessageLookupByLibrary.simpleMessage("SIGNIFICATION DES NOMBRES DU GRAPHIQUE")
-  };
+  static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "bieu_do_ngay_sinh": MessageLookupByLibrary.simpleMessage(
+            "Tableau des dates de naissance"),
+        "chon_ngon_ngu":
+            MessageLookupByLibrary.simpleMessage("Sélectionnez la langue"),
+        "con_so_chu_dao_cua_ban_la":
+            MessageLookupByLibrary.simpleMessage("Votre numéro fatidique"),
+        "de_xuat":
+            MessageLookupByLibrary.simpleMessage("Développement proposé"),
+        "giai_thich": MessageLookupByLibrary.simpleMessage("Expliquez"),
+        "gioi_thieu": MessageLookupByLibrary.simpleMessage("Introl"),
+        "hay_chon_ngay_sinh_cua_ban": MessageLookupByLibrary.simpleMessage(
+            "Veuillez sélectionner votre date de naissance"),
+        "home": MessageLookupByLibrary.simpleMessage("Maison"),
+        "introl_1": MessageLookupByLibrary.simpleMessage(
+            "La numérologie est basée sur votre nom et votre date de naissance pour vous prédire de manière complète. Avec la numérologie, voyez comment les nombres importants affectent votre vie."),
+        "introl_2": MessageLookupByLibrary.simpleMessage(
+            "L\'application analysera votre nom et votre date de naissance pour afficher les résultats en un clin d\'œil! En ​​savoir plus sur les numéros de naissance ou les lignes de vie (exprimez votre personnalité et vos perspectives dans la vie). Voyons ce que ces chiffres disent sur toi"),
+        "khuyet_diem": MessageLookupByLibrary.simpleMessage("Faiblesse"),
+        "moc_cuoc_doi":
+            MessageLookupByLibrary.simpleMessage("Jalons de la vie"),
+        "muc_dich_cuoc_song":
+            MessageLookupByLibrary.simpleMessage("Objectifs de vie"),
+        "nghe_nghiep_phu_hop":
+            MessageLookupByLibrary.simpleMessage("Occupation convenable"),
+        "ngon_ngu": MessageLookupByLibrary.simpleMessage("Langue"),
+        "noi_bat": MessageLookupByLibrary.simpleMessage("En vedette"),
+        "phat_trien_theo_so":
+            MessageLookupByLibrary.simpleMessage("Croissance"),
+        "so": MessageLookupByLibrary.simpleMessage("Number"),
+        "than_so_hoc_la_gi": MessageLookupByLibrary.simpleMessage(
+            "QU\'EST-CE QUE LA Numérologie?"),
+        "thiet_lap_ngay_sinh":
+            MessageLookupByLibrary.simpleMessage("set date of birth"),
+        "tong_quan": MessageLookupByLibrary.simpleMessage("Présentation"),
+        "tu_vi_hom_nay":
+            MessageLookupByLibrary.simpleMessage("Horoscope today"),
+        "tuoi": MessageLookupByLibrary.simpleMessage("Ans"),
+        "uu_diem": MessageLookupByLibrary.simpleMessage("Avantages"),
+        "xem": MessageLookupByLibrary.simpleMessage("Soumettre"),
+        "xem_ngay_sinh": MessageLookupByLibrary.simpleMessage("Accueil"),
+        "y_nghia_cac_con_so_tren_bieu_do": MessageLookupByLibrary.simpleMessage(
+            "SIGNIFICATION DES NOMBRES DU GRAPHIQUE")
+      };
 }
