@@ -7,7 +7,8 @@
 // ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
-// ignore_for_file:unused_import, file_names
+// ignore_for_file:unused_import, file_names, avoid_escaping_inner_quotes
+// ignore_for_file:unnecessary_string_interpolations, unnecessary_string_escapes
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -20,33 +21,45 @@ class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'hi';
 
   final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "bieu_do_ngay_sinh" : MessageLookupByLibrary.simpleMessage("जन्म तिथि चार्ट"),
-    "chon_ngon_ngu" : MessageLookupByLibrary.simpleMessage("भाषा चुनें"),
-    "con_so_chu_dao_cua_ban_la" : MessageLookupByLibrary.simpleMessage("आपका भाग्य नंबर"),
-    "de_xuat" : MessageLookupByLibrary.simpleMessage("प्रस्तावित विकास"),
-    "giai_thich" : MessageLookupByLibrary.simpleMessage("समझाएं"),
-    "gioi_thieu" : MessageLookupByLibrary.simpleMessage("परिचय कराना"),
-    "hay_chon_ngay_sinh_cua_ban" : MessageLookupByLibrary.simpleMessage("कृपया अपनी जन्मतिथि चुनें"),
-    "home" : MessageLookupByLibrary.simpleMessage("घर"),
-    "introl_1" : MessageLookupByLibrary.simpleMessage("न्यूमेरोलॉजी आपके नाम और जन्म की तारीख पर आधारित है ताकि आप बड़े पैमाने पर भविष्यवाणी कर सकें। न्यूमेरोलॉजी के साथ, देखें कि महत्वपूर्ण संख्या आपके जीवन को कैसे प्रभावित करती है।"),
-    "introl_2" : MessageLookupByLibrary.simpleMessage("ऐप पलक झपकते ही परिणाम दिखाने के लिए आपके नाम और जन्म तिथि का विश्लेषण करेगा! जन्म संख्या या जीवन रेखा के बारे में जानें (जीवन में अपने व्यक्तित्व और दृष्टिकोण को व्यक्त करें)। आइए देखें कि ये संख्याएँ क्या कहती हैं। आप"),
-    "khuyet_diem" : MessageLookupByLibrary.simpleMessage("कमजोरी"),
-    "moc_cuoc_doi" : MessageLookupByLibrary.simpleMessage("लाइफ मील के पत्थर"),
-    "muc_dich_cuoc_song" : MessageLookupByLibrary.simpleMessage("जीवन लक्ष्य"),
-    "nghe_nghiep_phu_hop" : MessageLookupByLibrary.simpleMessage("उपयुक्त व्यवसाय"),
-    "ngon_ngu" : MessageLookupByLibrary.simpleMessage("भाषा"),
-    "noi_bat" : MessageLookupByLibrary.simpleMessage("फीचर्ड"),
-    "phat_trien_theo_so" : MessageLookupByLibrary.simpleMessage("विकास"),
-    "so" : MessageLookupByLibrary.simpleMessage("नंबर"),
-    "than_so_hoc_la_gi" : MessageLookupByLibrary.simpleMessage("संख्या विज्ञान क्या है?"),
-    "thiet_lap_ngay_sinh" : MessageLookupByLibrary.simpleMessage("set date of birth"),
-    "tong_quan" : MessageLookupByLibrary.simpleMessage("ओवरव्यू"),
-    "tu_vi_hom_nay" : MessageLookupByLibrary.simpleMessage("Horoscope today"),
-    "tuoi" : MessageLookupByLibrary.simpleMessage("वर्ष पुराना"),
-    "uu_diem" : MessageLookupByLibrary.simpleMessage("लाभ"),
-    "xem" : MessageLookupByLibrary.simpleMessage("सबमिट करें"),
-    "xem_ngay_sinh" : MessageLookupByLibrary.simpleMessage("होम"),
-    "y_nghia_cac_con_so_tren_bieu_do" : MessageLookupByLibrary.simpleMessage(" ग्राफ़ के नामकरण")
-  };
+  static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "bieu_do_ngay_sinh":
+            MessageLookupByLibrary.simpleMessage("जन्म तिथि चार्ट"),
+        "chon_ngon_ngu": MessageLookupByLibrary.simpleMessage("भाषा चुनें"),
+        "con_so_chu_dao_cua_ban_la":
+            MessageLookupByLibrary.simpleMessage("आपका भाग्य नंबर"),
+        "de_xuat": MessageLookupByLibrary.simpleMessage("प्रस्तावित विकास"),
+        "giai_thich": MessageLookupByLibrary.simpleMessage("समझाएं"),
+        "gioi_thieu": MessageLookupByLibrary.simpleMessage("परिचय कराना"),
+        "hay_chon_ngay_sinh_cua_ban":
+            MessageLookupByLibrary.simpleMessage("कृपया अपनी जन्मतिथि चुनें"),
+        "home": MessageLookupByLibrary.simpleMessage("घर"),
+        "introl_1": MessageLookupByLibrary.simpleMessage(
+            "न्यूमेरोलॉजी आपके नाम और जन्म की तारीख पर आधारित है ताकि आप बड़े पैमाने पर भविष्यवाणी कर सकें। न्यूमेरोलॉजी के साथ, देखें कि महत्वपूर्ण संख्या आपके जीवन को कैसे प्रभावित करती है।"),
+        "introl_2": MessageLookupByLibrary.simpleMessage(
+            "ऐप पलक झपकते ही परिणाम दिखाने के लिए आपके नाम और जन्म तिथि का विश्लेषण करेगा! जन्म संख्या या जीवन रेखा के बारे में जानें (जीवन में अपने व्यक्तित्व और दृष्टिकोण को व्यक्त करें)। आइए देखें कि ये संख्याएँ क्या कहती हैं। आप"),
+        "khuyet_diem": MessageLookupByLibrary.simpleMessage("कमजोरी"),
+        "moc_cuoc_doi":
+            MessageLookupByLibrary.simpleMessage("लाइफ मील के पत्थर"),
+        "muc_dich_cuoc_song":
+            MessageLookupByLibrary.simpleMessage("जीवन लक्ष्य"),
+        "nghe_nghiep_phu_hop":
+            MessageLookupByLibrary.simpleMessage("उपयुक्त व्यवसाय"),
+        "ngon_ngu": MessageLookupByLibrary.simpleMessage("भाषा"),
+        "noi_bat": MessageLookupByLibrary.simpleMessage("फीचर्ड"),
+        "phat_trien_theo_so": MessageLookupByLibrary.simpleMessage("विकास"),
+        "so": MessageLookupByLibrary.simpleMessage("नंबर"),
+        "than_so_hoc_la_gi":
+            MessageLookupByLibrary.simpleMessage("संख्या विज्ञान क्या है?"),
+        "thiet_lap_ngay_sinh":
+            MessageLookupByLibrary.simpleMessage("set date of birth"),
+        "tong_quan": MessageLookupByLibrary.simpleMessage("ओवरव्यू"),
+        "tu_vi_hom_nay":
+            MessageLookupByLibrary.simpleMessage("Horoscope today"),
+        "tuoi": MessageLookupByLibrary.simpleMessage("वर्ष पुराना"),
+        "uu_diem": MessageLookupByLibrary.simpleMessage("लाभ"),
+        "xem": MessageLookupByLibrary.simpleMessage("सबमिट करें"),
+        "xem_ngay_sinh": MessageLookupByLibrary.simpleMessage("होम"),
+        "y_nghia_cac_con_so_tren_bieu_do":
+            MessageLookupByLibrary.simpleMessage(" ग्राफ़ के नामकरण")
+      };
 }
