@@ -7,7 +7,8 @@
 // ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
-// ignore_for_file:unused_import, file_names
+// ignore_for_file:unused_import, file_names, avoid_escaping_inner_quotes
+// ignore_for_file:unnecessary_string_interpolations, unnecessary_string_escapes
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -20,33 +21,47 @@ class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'in';
 
   final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "bieu_do_ngay_sinh" : MessageLookupByLibrary.simpleMessage("Bagan tanggal lahir"),
-    "chon_ngon_ngu" : MessageLookupByLibrary.simpleMessage("Pilih bahasa"),
-    "con_so_chu_dao_cua_ban_la" : MessageLookupByLibrary.simpleMessage("Nomor penting Anda"),
-    "de_xuat" : MessageLookupByLibrary.simpleMessage("Pengembangan yang diusulkan"),
-    "giai_thich" : MessageLookupByLibrary.simpleMessage("Jelaskan"),
-    "gioi_thieu" : MessageLookupByLibrary.simpleMessage("Introl"),
-    "hay_chon_ngay_sinh_cua_ban" : MessageLookupByLibrary.simpleMessage("Silakan pilih tanggal lahir Anda"),
-    "home" : MessageLookupByLibrary.simpleMessage("Rumah"),
-    "introl_1" : MessageLookupByLibrary.simpleMessage("Numerologi didasarkan pada nama dan tanggal lahir Anda untuk memprediksi Anda secara komprehensif. Dengan Numerologi, lihat seberapa penting angka mempengaruhi hidup Anda."),
-    "introl_2" : MessageLookupByLibrary.simpleMessage("Aplikasi akan menganalisis nama dan tanggal lahir Anda untuk menunjukkan hasil dalam sekejap mata! Pelajari tentang nomor kelahiran atau garis kehidupan (ekspresikan kepribadian dan pandangan hidup Anda). Mari kita lihat apa yang dikatakan angka-angka ini kamu"),
-    "khuyet_diem" : MessageLookupByLibrary.simpleMessage("Kelemahan"),
-    "moc_cuoc_doi" : MessageLookupByLibrary.simpleMessage("Tonggak kehidupan"),
-    "muc_dich_cuoc_song" : MessageLookupByLibrary.simpleMessage("Tujuan hidup"),
-    "nghe_nghiep_phu_hop" : MessageLookupByLibrary.simpleMessage("Pekerjaan yang sesuai"),
-    "ngon_ngu" : MessageLookupByLibrary.simpleMessage("Bahasa"),
-    "noi_bat" : MessageLookupByLibrary.simpleMessage("Unggulan"),
-    "phat_trien_theo_so" : MessageLookupByLibrary.simpleMessage("Pertumbuhan"),
-    "so" : MessageLookupByLibrary.simpleMessage("Nomor"),
-    "than_so_hoc_la_gi" : MessageLookupByLibrary.simpleMessage("APA ITU Numerologi?"),
-    "thiet_lap_ngay_sinh" : MessageLookupByLibrary.simpleMessage("set date of birth"),
-    "tong_quan" : MessageLookupByLibrary.simpleMessage("Ringkasan"),
-    "tu_vi_hom_nay" : MessageLookupByLibrary.simpleMessage("Horoscope today"),
-    "tuoi" : MessageLookupByLibrary.simpleMessage("Tahun tua"),
-    "uu_diem" : MessageLookupByLibrary.simpleMessage("Keuntungan"),
-    "xem" : MessageLookupByLibrary.simpleMessage("Kirim"),
-    "xem_ngay_sinh" : MessageLookupByLibrary.simpleMessage("Rumah"),
-    "y_nghia_cac_con_so_tren_bieu_do" : MessageLookupByLibrary.simpleMessage("BERARTI ANGKA")
-  };
+  static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "bieu_do_ngay_sinh":
+            MessageLookupByLibrary.simpleMessage("Bagan tanggal lahir"),
+        "chon_ngon_ngu": MessageLookupByLibrary.simpleMessage("Pilih bahasa"),
+        "con_so_chu_dao_cua_ban_la":
+            MessageLookupByLibrary.simpleMessage("Nomor penting Anda"),
+        "de_xuat":
+            MessageLookupByLibrary.simpleMessage("Pengembangan yang diusulkan"),
+        "giai_thich": MessageLookupByLibrary.simpleMessage("Jelaskan"),
+        "gioi_thieu": MessageLookupByLibrary.simpleMessage("Introl"),
+        "hay_chon_ngay_sinh_cua_ban": MessageLookupByLibrary.simpleMessage(
+            "Silakan pilih tanggal lahir Anda"),
+        "home": MessageLookupByLibrary.simpleMessage("Rumah"),
+        "introl_1": MessageLookupByLibrary.simpleMessage(
+            "Numerologi didasarkan pada nama dan tanggal lahir Anda untuk memprediksi Anda secara komprehensif. Dengan Numerologi, lihat seberapa penting angka mempengaruhi hidup Anda."),
+        "introl_2": MessageLookupByLibrary.simpleMessage(
+            "Aplikasi akan menganalisis nama dan tanggal lahir Anda untuk menunjukkan hasil dalam sekejap mata! Pelajari tentang nomor kelahiran atau garis kehidupan (ekspresikan kepribadian dan pandangan hidup Anda). Mari kita lihat apa yang dikatakan angka-angka ini kamu"),
+        "khuyet_diem": MessageLookupByLibrary.simpleMessage("Kelemahan"),
+        "moc_cuoc_doi":
+            MessageLookupByLibrary.simpleMessage("Tonggak kehidupan"),
+        "muc_dich_cuoc_song":
+            MessageLookupByLibrary.simpleMessage("Tujuan hidup"),
+        "nghe_nghiep_phu_hop":
+            MessageLookupByLibrary.simpleMessage("Pekerjaan yang sesuai"),
+        "ngon_ngu": MessageLookupByLibrary.simpleMessage("Bahasa"),
+        "noi_bat": MessageLookupByLibrary.simpleMessage("Unggulan"),
+        "phat_trien_theo_so":
+            MessageLookupByLibrary.simpleMessage("Pertumbuhan"),
+        "so": MessageLookupByLibrary.simpleMessage("Nomor"),
+        "than_so_hoc_la_gi":
+            MessageLookupByLibrary.simpleMessage("APA ITU Numerologi?"),
+        "thiet_lap_ngay_sinh":
+            MessageLookupByLibrary.simpleMessage("set date of birth"),
+        "tong_quan": MessageLookupByLibrary.simpleMessage("Ringkasan"),
+        "tu_vi_hom_nay":
+            MessageLookupByLibrary.simpleMessage("Horoscope today"),
+        "tuoi": MessageLookupByLibrary.simpleMessage("Tahun tua"),
+        "uu_diem": MessageLookupByLibrary.simpleMessage("Keuntungan"),
+        "xem": MessageLookupByLibrary.simpleMessage("Kirim"),
+        "xem_ngay_sinh": MessageLookupByLibrary.simpleMessage("Rumah"),
+        "y_nghia_cac_con_so_tren_bieu_do":
+            MessageLookupByLibrary.simpleMessage("BERARTI ANGKA")
+      };
 }
