@@ -7,7 +7,8 @@
 // ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
-// ignore_for_file:unused_import, file_names
+// ignore_for_file:unused_import, file_names, avoid_escaping_inner_quotes
+// ignore_for_file:unnecessary_string_interpolations, unnecessary_string_escapes
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -20,33 +21,45 @@ class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'lo';
 
   final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "bieu_do_ngay_sinh" : MessageLookupByLibrary.simpleMessage("ຕາຕະລາງວັນເດືອນປີເກີດ"),
-    "chon_ngon_ngu" : MessageLookupByLibrary.simpleMessage("ເລືອກພາສາ"),
-    "con_so_chu_dao_cua_ban_la" : MessageLookupByLibrary.simpleMessage("ເລກໂຊກຊະຕາຂອງທ່ານ"),
-    "de_xuat" : MessageLookupByLibrary.simpleMessage("ການພັດທະນາທີ່ສະ ເໜີ"),
-    "giai_thich" : MessageLookupByLibrary.simpleMessage("ອະທິບາຍ"),
-    "gioi_thieu" : MessageLookupByLibrary.simpleMessage("Introl"),
-    "hay_chon_ngay_sinh_cua_ban" : MessageLookupByLibrary.simpleMessage("ກະລຸນາເລືອກວັນເດືອນປີເກີດຂອງທ່ານ"),
-    "home" : MessageLookupByLibrary.simpleMessage("ບ້ານ"),
-    "introl_1" : MessageLookupByLibrary.simpleMessage("Numerology ແມ່ນອີງໃສ່ຊື່ແລະວັນເດືອນປີເກີດຂອງທ່ານເພື່ອຄາດເດົາທ່ານໃຫ້ສົມບູນແບບກັບ Numerology, ເບິ່ງວ່າຕົວເລກທີ່ ສຳ ຄັນມີຜົນກະທົບຕໍ່ຊີວິດຂອງທ່ານແນວໃດ."),
-    "introl_2" : MessageLookupByLibrary.simpleMessage("ແອັບ will ຈະວິເຄາະຊື່ແລະວັນເດືອນປີເກີດຂອງທ່ານເພື່ອສະແດງໃຫ້ເຫັນຜົນໃນກະພິບ! ຮຽນຮູ້ກ່ຽວກັບຕົວເລກຂອງການເກີດຫຼືສາຍຊີວິດ (ສະແດງເຖິງບຸກຄະລິກແລະມຸມມອງຂອງທ່ານໃນຊີວິດ) ລອງເບິ່ງວ່າຕົວເລກເຫລົ່ານີ້ເວົ້າກ່ຽວກັບຫຍັງ ເຈົ້າ "),
-    "khuyet_diem" : MessageLookupByLibrary.simpleMessage("ຄວາມອ່ອນແອ"),
-    "moc_cuoc_doi" : MessageLookupByLibrary.simpleMessage("ຈຸດ ສຳ ຄັນຂອງຊີວິດ"),
-    "muc_dich_cuoc_song" : MessageLookupByLibrary.simpleMessage("ເປົ້າ ໝາຍ ຊີວິດ"),
-    "nghe_nghiep_phu_hop" : MessageLookupByLibrary.simpleMessage("ອາຊີບທີ່ ເໝາະ ສົມ"),
-    "ngon_ngu" : MessageLookupByLibrary.simpleMessage("ພາສາ"),
-    "noi_bat" : MessageLookupByLibrary.simpleMessage("ເດັ່ນ"),
-    "phat_trien_theo_so" : MessageLookupByLibrary.simpleMessage("ການເຕີບໂຕ"),
-    "so" : MessageLookupByLibrary.simpleMessage("ໝາຍ ເລກ"),
-    "than_so_hoc_la_gi" : MessageLookupByLibrary.simpleMessage("Numerology ແມ່ນຫຍັງ?"),
-    "thiet_lap_ngay_sinh" : MessageLookupByLibrary.simpleMessage("set date of birth"),
-    "tong_quan" : MessageLookupByLibrary.simpleMessage("ພາບລວມ"),
-    "tu_vi_hom_nay" : MessageLookupByLibrary.simpleMessage("Horoscope today"),
-    "tuoi" : MessageLookupByLibrary.simpleMessage("ປີເກົ່າ"),
-    "uu_diem" : MessageLookupByLibrary.simpleMessage("ຂໍ້ໄດ້ປຽບ"),
-    "xem" : MessageLookupByLibrary.simpleMessage("ສົ່ງ"),
-    "xem_ngay_sinh" : MessageLookupByLibrary.simpleMessage("ບ້ານ"),
-    "y_nghia_cac_con_so_tren_bieu_do" : MessageLookupByLibrary.simpleMessage("ໝາຍ ຄວາມວ່າບັນດາ ກຳ ມະກອນຂອງ GRAPH")
-  };
+  static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "bieu_do_ngay_sinh":
+            MessageLookupByLibrary.simpleMessage("ຕາຕະລາງວັນເດືອນປີເກີດ"),
+        "chon_ngon_ngu": MessageLookupByLibrary.simpleMessage("ເລືອກພາສາ"),
+        "con_so_chu_dao_cua_ban_la":
+            MessageLookupByLibrary.simpleMessage("ເລກໂຊກຊະຕາຂອງທ່ານ"),
+        "de_xuat": MessageLookupByLibrary.simpleMessage("ການພັດທະນາທີ່ສະ ເໜີ"),
+        "giai_thich": MessageLookupByLibrary.simpleMessage("ອະທິບາຍ"),
+        "gioi_thieu": MessageLookupByLibrary.simpleMessage("Introl"),
+        "hay_chon_ngay_sinh_cua_ban": MessageLookupByLibrary.simpleMessage(
+            "ກະລຸນາເລືອກວັນເດືອນປີເກີດຂອງທ່ານ"),
+        "home": MessageLookupByLibrary.simpleMessage("ບ້ານ"),
+        "introl_1": MessageLookupByLibrary.simpleMessage(
+            "Numerology ແມ່ນອີງໃສ່ຊື່ແລະວັນເດືອນປີເກີດຂອງທ່ານເພື່ອຄາດເດົາທ່ານໃຫ້ສົມບູນແບບກັບ Numerology, ເບິ່ງວ່າຕົວເລກທີ່ ສຳ ຄັນມີຜົນກະທົບຕໍ່ຊີວິດຂອງທ່ານແນວໃດ."),
+        "introl_2": MessageLookupByLibrary.simpleMessage(
+            "ແອັບ will ຈະວິເຄາະຊື່ແລະວັນເດືອນປີເກີດຂອງທ່ານເພື່ອສະແດງໃຫ້ເຫັນຜົນໃນກະພິບ! ຮຽນຮູ້ກ່ຽວກັບຕົວເລກຂອງການເກີດຫຼືສາຍຊີວິດ (ສະແດງເຖິງບຸກຄະລິກແລະມຸມມອງຂອງທ່ານໃນຊີວິດ) ລອງເບິ່ງວ່າຕົວເລກເຫລົ່ານີ້ເວົ້າກ່ຽວກັບຫຍັງ ເຈົ້າ "),
+        "khuyet_diem": MessageLookupByLibrary.simpleMessage("ຄວາມອ່ອນແອ"),
+        "moc_cuoc_doi":
+            MessageLookupByLibrary.simpleMessage("ຈຸດ ສຳ ຄັນຂອງຊີວິດ"),
+        "muc_dich_cuoc_song":
+            MessageLookupByLibrary.simpleMessage("ເປົ້າ ໝາຍ ຊີວິດ"),
+        "nghe_nghiep_phu_hop":
+            MessageLookupByLibrary.simpleMessage("ອາຊີບທີ່ ເໝາະ ສົມ"),
+        "ngon_ngu": MessageLookupByLibrary.simpleMessage("ພາສາ"),
+        "noi_bat": MessageLookupByLibrary.simpleMessage("ເດັ່ນ"),
+        "phat_trien_theo_so": MessageLookupByLibrary.simpleMessage("ການເຕີບໂຕ"),
+        "so": MessageLookupByLibrary.simpleMessage("ໝາຍ ເລກ"),
+        "than_so_hoc_la_gi":
+            MessageLookupByLibrary.simpleMessage("Numerology ແມ່ນຫຍັງ?"),
+        "thiet_lap_ngay_sinh":
+            MessageLookupByLibrary.simpleMessage("set date of birth"),
+        "tong_quan": MessageLookupByLibrary.simpleMessage("ພາບລວມ"),
+        "tu_vi_hom_nay":
+            MessageLookupByLibrary.simpleMessage("Horoscope today"),
+        "tuoi": MessageLookupByLibrary.simpleMessage("ປີເກົ່າ"),
+        "uu_diem": MessageLookupByLibrary.simpleMessage("ຂໍ້ໄດ້ປຽບ"),
+        "xem": MessageLookupByLibrary.simpleMessage("ສົ່ງ"),
+        "xem_ngay_sinh": MessageLookupByLibrary.simpleMessage("ບ້ານ"),
+        "y_nghia_cac_con_so_tren_bieu_do": MessageLookupByLibrary.simpleMessage(
+            "ໝາຍ ຄວາມວ່າບັນດາ ກຳ ມະກອນຂອງ GRAPH")
+      };
 }
