@@ -7,7 +7,8 @@
 // ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
-// ignore_for_file:unused_import, file_names
+// ignore_for_file:unused_import, file_names, avoid_escaping_inner_quotes
+// ignore_for_file:unnecessary_string_interpolations, unnecessary_string_escapes
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -20,33 +21,47 @@ class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'pt';
 
   final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "bieu_do_ngay_sinh" : MessageLookupByLibrary.simpleMessage("Tabela de data de nascimento"),
-    "chon_ngon_ngu" : MessageLookupByLibrary.simpleMessage("Selecione o idioma"),
-    "con_so_chu_dao_cua_ban_la" : MessageLookupByLibrary.simpleMessage("Seu número fatídico"),
-    "de_xuat" : MessageLookupByLibrary.simpleMessage("Desenvolvimento proposto"),
-    "giai_thich" : MessageLookupByLibrary.simpleMessage("Explique"),
-    "gioi_thieu" : MessageLookupByLibrary.simpleMessage("Introl"),
-    "hay_chon_ngay_sinh_cua_ban" : MessageLookupByLibrary.simpleMessage("Selecione sua data de nascimento"),
-    "home" : MessageLookupByLibrary.simpleMessage("Casa"),
-    "introl_1" : MessageLookupByLibrary.simpleMessage("A numerologia se baseia em seu nome e data de nascimento para prever você de forma abrangente. Com a numerologia, veja como números importantes afetam sua vida."),
-    "introl_2" : MessageLookupByLibrary.simpleMessage("O aplicativo analisará seu nome e data de nascimento para mostrar os resultados em um piscar de olhos! Saiba mais sobre números de nascimento ou linhas de vida (expresse sua personalidade e perspectiva de vida). Vamos ver o que esses números dizem sobre vocês"),
-    "khuyet_diem" : MessageLookupByLibrary.simpleMessage("Fraqueza"),
-    "moc_cuoc_doi" : MessageLookupByLibrary.simpleMessage("Marcos da vida"),
-    "muc_dich_cuoc_song" : MessageLookupByLibrary.simpleMessage("Objetivos de vida"),
-    "nghe_nghiep_phu_hop" : MessageLookupByLibrary.simpleMessage("Ocupação adequada"),
-    "ngon_ngu" : MessageLookupByLibrary.simpleMessage("Idioma"),
-    "noi_bat" : MessageLookupByLibrary.simpleMessage("Apresentado"),
-    "phat_trien_theo_so" : MessageLookupByLibrary.simpleMessage("Crescimento"),
-    "so" : MessageLookupByLibrary.simpleMessage("Número"),
-    "than_so_hoc_la_gi" : MessageLookupByLibrary.simpleMessage("O QUE É Numerologia?"),
-    "thiet_lap_ngay_sinh" : MessageLookupByLibrary.simpleMessage("set date of birth"),
-    "tong_quan" : MessageLookupByLibrary.simpleMessage("Visão geral"),
-    "tu_vi_hom_nay" : MessageLookupByLibrary.simpleMessage("Horoscope today"),
-    "tuoi" : MessageLookupByLibrary.simpleMessage("Anos"),
-    "uu_diem" : MessageLookupByLibrary.simpleMessage("Vantagens"),
-    "xem" : MessageLookupByLibrary.simpleMessage("Enviar"),
-    "xem_ngay_sinh" : MessageLookupByLibrary.simpleMessage("Home"),
-    "y_nghia_cac_con_so_tren_bieu_do" : MessageLookupByLibrary.simpleMessage("SIGNIFICADO OS NÚMEROS DO GRÁFICO")
-  };
+  static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "bieu_do_ngay_sinh": MessageLookupByLibrary.simpleMessage(
+            "Tabela de data de nascimento"),
+        "chon_ngon_ngu":
+            MessageLookupByLibrary.simpleMessage("Selecione o idioma"),
+        "con_so_chu_dao_cua_ban_la":
+            MessageLookupByLibrary.simpleMessage("Seu número fatídico"),
+        "de_xuat":
+            MessageLookupByLibrary.simpleMessage("Desenvolvimento proposto"),
+        "giai_thich": MessageLookupByLibrary.simpleMessage("Explique"),
+        "gioi_thieu": MessageLookupByLibrary.simpleMessage("Introl"),
+        "hay_chon_ngay_sinh_cua_ban": MessageLookupByLibrary.simpleMessage(
+            "Selecione sua data de nascimento"),
+        "home": MessageLookupByLibrary.simpleMessage("Casa"),
+        "introl_1": MessageLookupByLibrary.simpleMessage(
+            "A numerologia se baseia em seu nome e data de nascimento para prever você de forma abrangente. Com a numerologia, veja como números importantes afetam sua vida."),
+        "introl_2": MessageLookupByLibrary.simpleMessage(
+            "O aplicativo analisará seu nome e data de nascimento para mostrar os resultados em um piscar de olhos! Saiba mais sobre números de nascimento ou linhas de vida (expresse sua personalidade e perspectiva de vida). Vamos ver o que esses números dizem sobre vocês"),
+        "khuyet_diem": MessageLookupByLibrary.simpleMessage("Fraqueza"),
+        "moc_cuoc_doi": MessageLookupByLibrary.simpleMessage("Marcos da vida"),
+        "muc_dich_cuoc_song":
+            MessageLookupByLibrary.simpleMessage("Objetivos de vida"),
+        "nghe_nghiep_phu_hop":
+            MessageLookupByLibrary.simpleMessage("Ocupação adequada"),
+        "ngon_ngu": MessageLookupByLibrary.simpleMessage("Idioma"),
+        "noi_bat": MessageLookupByLibrary.simpleMessage("Apresentado"),
+        "phat_trien_theo_so":
+            MessageLookupByLibrary.simpleMessage("Crescimento"),
+        "so": MessageLookupByLibrary.simpleMessage("Número"),
+        "than_so_hoc_la_gi":
+            MessageLookupByLibrary.simpleMessage("O QUE É Numerologia?"),
+        "thiet_lap_ngay_sinh":
+            MessageLookupByLibrary.simpleMessage("set date of birth"),
+        "tong_quan": MessageLookupByLibrary.simpleMessage("Visão geral"),
+        "tu_vi_hom_nay":
+            MessageLookupByLibrary.simpleMessage("Horoscope today"),
+        "tuoi": MessageLookupByLibrary.simpleMessage("Anos"),
+        "uu_diem": MessageLookupByLibrary.simpleMessage("Vantagens"),
+        "xem": MessageLookupByLibrary.simpleMessage("Enviar"),
+        "xem_ngay_sinh": MessageLookupByLibrary.simpleMessage("Home"),
+        "y_nghia_cac_con_so_tren_bieu_do": MessageLookupByLibrary.simpleMessage(
+            "SIGNIFICADO OS NÚMEROS DO GRÁFICO")
+      };
 }
