@@ -7,7 +7,8 @@
 // ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
-// ignore_for_file:unused_import, file_names
+// ignore_for_file:unused_import, file_names, avoid_escaping_inner_quotes
+// ignore_for_file:unnecessary_string_interpolations, unnecessary_string_escapes
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -20,32 +21,44 @@ class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'ru';
 
   final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "bieu_do_ngay_sinh" : MessageLookupByLibrary.simpleMessage("Таблица дат рождения"),
-    "chon_ngon_ngu" : MessageLookupByLibrary.simpleMessage("Выбрать язык"),
-    "con_so_chu_dao_cua_ban_la" : MessageLookupByLibrary.simpleMessage("Твой роковой номер"),
-    "de_xuat" : MessageLookupByLibrary.simpleMessage("Предлагаемая разработка"),
-    "giai_thich" : MessageLookupByLibrary.simpleMessage("Объясни"),
-    "gioi_thieu" : MessageLookupByLibrary.simpleMessage("Introl"),
-    "hay_chon_ngay_sinh_cua_ban" : MessageLookupByLibrary.simpleMessage("Пожалуйста, выберите дату вашего рождения"),
-    "home" : MessageLookupByLibrary.simpleMessage("Дом"),
-    "introl_1" : MessageLookupByLibrary.simpleMessage("Нумерология основана на вашем имени и дате рождения, чтобы предсказать вас всесторонне. С помощью нумерологии вы увидите, как важные числа влияют на вашу жизнь."),
-    "introl_2" : MessageLookupByLibrary.simpleMessage("Приложение проанализирует ваше имя и дату рождения, чтобы мгновенно отобразить результаты! Узнайте о числах рождения или жизненных линиях (выражайте вашу личность и мировоззрение). Давайте посмотрим, что говорят эти числа ты"),
-    "moc_cuoc_doi" : MessageLookupByLibrary.simpleMessage("Жизненные вехи"),
-    "muc_dich_cuoc_song" : MessageLookupByLibrary.simpleMessage("Жизненные цели"),
-    "nghe_nghiep_phu_hop" : MessageLookupByLibrary.simpleMessage("Подходящее занятие"),
-    "ngon_ngu" : MessageLookupByLibrary.simpleMessage("Язык"),
-    "noi_bat" : MessageLookupByLibrary.simpleMessage("Рекомендуемое"),
-    "phat_trien_theo_so" : MessageLookupByLibrary.simpleMessage("прирост"),
-    "so" : MessageLookupByLibrary.simpleMessage("Число"),
-    "than_so_hoc_la_gi" : MessageLookupByLibrary.simpleMessage("ЧТО ТАКОЕ нумерология?"),
-    "thiet_lap_ngay_sinh" : MessageLookupByLibrary.simpleMessage("set date of birth"),
-    "tong_quan" : MessageLookupByLibrary.simpleMessage("Обзор"),
-    "tu_vi_hom_nay" : MessageLookupByLibrary.simpleMessage("Horoscope today"),
-    "tuoi" : MessageLookupByLibrary.simpleMessage("Годовалый"),
-    "uu_diem" : MessageLookupByLibrary.simpleMessage("Преимущества"),
-    "xem" : MessageLookupByLibrary.simpleMessage("Отправить"),
-    "xem_ngay_sinh" : MessageLookupByLibrary.simpleMessage("Дом"),
-    "y_nghia_cac_con_so_tren_bieu_do" : MessageLookupByLibrary.simpleMessage("ЗНАЧЕНИЕ ЧИСЛОВ НА ГРАФИКЕ")
-  };
+  static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "bieu_do_ngay_sinh":
+            MessageLookupByLibrary.simpleMessage("Таблица дат рождения"),
+        "chon_ngon_ngu": MessageLookupByLibrary.simpleMessage("Выбрать язык"),
+        "con_so_chu_dao_cua_ban_la":
+            MessageLookupByLibrary.simpleMessage("Твой роковой номер"),
+        "de_xuat":
+            MessageLookupByLibrary.simpleMessage("Предлагаемая разработка"),
+        "giai_thich": MessageLookupByLibrary.simpleMessage("Объясни"),
+        "gioi_thieu": MessageLookupByLibrary.simpleMessage("Introl"),
+        "hay_chon_ngay_sinh_cua_ban": MessageLookupByLibrary.simpleMessage(
+            "Пожалуйста, выберите дату вашего рождения"),
+        "home": MessageLookupByLibrary.simpleMessage("Дом"),
+        "introl_1": MessageLookupByLibrary.simpleMessage(
+            "Нумерология основана на вашем имени и дате рождения, чтобы предсказать вас всесторонне. С помощью нумерологии вы увидите, как важные числа влияют на вашу жизнь."),
+        "introl_2": MessageLookupByLibrary.simpleMessage(
+            "Приложение проанализирует ваше имя и дату рождения, чтобы мгновенно отобразить результаты! Узнайте о числах рождения или жизненных линиях (выражайте вашу личность и мировоззрение). Давайте посмотрим, что говорят эти числа ты"),
+        "moc_cuoc_doi": MessageLookupByLibrary.simpleMessage("Жизненные вехи"),
+        "muc_dich_cuoc_song":
+            MessageLookupByLibrary.simpleMessage("Жизненные цели"),
+        "nghe_nghiep_phu_hop":
+            MessageLookupByLibrary.simpleMessage("Подходящее занятие"),
+        "ngon_ngu": MessageLookupByLibrary.simpleMessage("Язык"),
+        "noi_bat": MessageLookupByLibrary.simpleMessage("Рекомендуемое"),
+        "phat_trien_theo_so": MessageLookupByLibrary.simpleMessage("прирост"),
+        "so": MessageLookupByLibrary.simpleMessage("Число"),
+        "than_so_hoc_la_gi":
+            MessageLookupByLibrary.simpleMessage("ЧТО ТАКОЕ нумерология?"),
+        "thiet_lap_ngay_sinh":
+            MessageLookupByLibrary.simpleMessage("set date of birth"),
+        "tong_quan": MessageLookupByLibrary.simpleMessage("Обзор"),
+        "tu_vi_hom_nay":
+            MessageLookupByLibrary.simpleMessage("Horoscope today"),
+        "tuoi": MessageLookupByLibrary.simpleMessage("Годовалый"),
+        "uu_diem": MessageLookupByLibrary.simpleMessage("Преимущества"),
+        "xem": MessageLookupByLibrary.simpleMessage("Отправить"),
+        "xem_ngay_sinh": MessageLookupByLibrary.simpleMessage("Дом"),
+        "y_nghia_cac_con_so_tren_bieu_do":
+            MessageLookupByLibrary.simpleMessage("ЗНАЧЕНИЕ ЧИСЛОВ НА ГРАФИКЕ")
+      };
 }
