@@ -7,7 +7,8 @@
 // ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
-// ignore_for_file:unused_import, file_names
+// ignore_for_file:unused_import, file_names, avoid_escaping_inner_quotes
+// ignore_for_file:unnecessary_string_interpolations, unnecessary_string_escapes
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -20,33 +21,44 @@ class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'vi';
 
   final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "bieu_do_ngay_sinh" : MessageLookupByLibrary.simpleMessage("Biểu đồ ngày sinh"),
-    "chon_ngon_ngu" : MessageLookupByLibrary.simpleMessage("Chọn ngôn ngữ"),
-    "con_so_chu_dao_cua_ban_la" : MessageLookupByLibrary.simpleMessage("Số chủ đạo của bạn là"),
-    "de_xuat" : MessageLookupByLibrary.simpleMessage("Đề xuất"),
-    "giai_thich" : MessageLookupByLibrary.simpleMessage("Giải thích"),
-    "gioi_thieu" : MessageLookupByLibrary.simpleMessage("Giới thiệu"),
-    "hay_chon_ngay_sinh_cua_ban" : MessageLookupByLibrary.simpleMessage("Hãy chọn ngày sinh của bạn"),
-    "home" : MessageLookupByLibrary.simpleMessage("THẦN SỐ HỌC"),
-    "introl_1" : MessageLookupByLibrary.simpleMessage("Thần số học sẽ dựa trên tên và ngày sinh của bạn để dự đoán về bạn một cách toàn diện. Với Thần số học toàn diện, hãy xem những con số quan trọng có ảnh hưởng như thế nào với cuộc sống của bạn."),
-    "introl_2" : MessageLookupByLibrary.simpleMessage("Ứng dụng sẽ phân tích tên và ngày sinh của bạn để đưa ra kết quả trong chớp mắt! Tìm hiểu về số sinh hoặc số đường đời (nói lên tính cách và quan điểm của bạn trong cuộc sống). Hãy xem những con số này nói điều gì về bạn"),
-    "khuyet_diem" : MessageLookupByLibrary.simpleMessage("Khuyết điểm"),
-    "moc_cuoc_doi" : MessageLookupByLibrary.simpleMessage("Mốc cuộc đời"),
-    "muc_dich_cuoc_song" : MessageLookupByLibrary.simpleMessage("Mục đích cuộc sống"),
-    "nghe_nghiep_phu_hop" : MessageLookupByLibrary.simpleMessage("Nghề nghiệp phù hợp"),
-    "ngon_ngu" : MessageLookupByLibrary.simpleMessage("Ngôn ngữ"),
-    "noi_bat" : MessageLookupByLibrary.simpleMessage("Nổi bật"),
-    "phat_trien_theo_so" : MessageLookupByLibrary.simpleMessage("Phát triển mạnh số"),
-    "so" : MessageLookupByLibrary.simpleMessage("Con số"),
-    "than_so_hoc_la_gi" : MessageLookupByLibrary.simpleMessage("THẦN SỐ HỌC LÀ GÌ ?"),
-    "thiet_lap_ngay_sinh" : MessageLookupByLibrary.simpleMessage("Thiết lập ngày sinh"),
-    "tong_quan" : MessageLookupByLibrary.simpleMessage("Tổng quan"),
-    "tu_vi_hom_nay" : MessageLookupByLibrary.simpleMessage("Tử vi hôm nay"),
-    "tuoi" : MessageLookupByLibrary.simpleMessage("Tuổi"),
-    "uu_diem" : MessageLookupByLibrary.simpleMessage("Ưu điểm"),
-    "xem" : MessageLookupByLibrary.simpleMessage("Xem"),
-    "xem_ngay_sinh" : MessageLookupByLibrary.simpleMessage("Xem ngày sinh"),
-    "y_nghia_cac_con_so_tren_bieu_do" : MessageLookupByLibrary.simpleMessage("MÝ NGHĨA CÁC CON SỐ TRÊN BIỂU ĐỒ")
-  };
+  static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "bieu_do_ngay_sinh":
+            MessageLookupByLibrary.simpleMessage("Biểu đồ ngày sinh"),
+        "chon_ngon_ngu": MessageLookupByLibrary.simpleMessage("Chọn ngôn ngữ"),
+        "con_so_chu_dao_cua_ban_la":
+            MessageLookupByLibrary.simpleMessage("Số chủ đạo của bạn là"),
+        "de_xuat": MessageLookupByLibrary.simpleMessage("Đề xuất"),
+        "giai_thich": MessageLookupByLibrary.simpleMessage("Giải thích"),
+        "gioi_thieu": MessageLookupByLibrary.simpleMessage("Giới thiệu"),
+        "hay_chon_ngay_sinh_cua_ban":
+            MessageLookupByLibrary.simpleMessage("Hãy chọn ngày sinh của bạn"),
+        "home": MessageLookupByLibrary.simpleMessage("THẦN SỐ HỌC"),
+        "introl_1": MessageLookupByLibrary.simpleMessage(
+            "Thần số học sẽ dựa trên tên và ngày sinh của bạn để dự đoán về bạn một cách toàn diện. Với Thần số học toàn diện, hãy xem những con số quan trọng có ảnh hưởng như thế nào với cuộc sống của bạn."),
+        "introl_2": MessageLookupByLibrary.simpleMessage(
+            "Ứng dụng sẽ phân tích tên và ngày sinh của bạn để đưa ra kết quả trong chớp mắt! Tìm hiểu về số sinh hoặc số đường đời (nói lên tính cách và quan điểm của bạn trong cuộc sống). Hãy xem những con số này nói điều gì về bạn"),
+        "khuyet_diem": MessageLookupByLibrary.simpleMessage("Khuyết điểm"),
+        "moc_cuoc_doi": MessageLookupByLibrary.simpleMessage("Mốc cuộc đời"),
+        "muc_dich_cuoc_song":
+            MessageLookupByLibrary.simpleMessage("Mục đích cuộc sống"),
+        "nghe_nghiep_phu_hop":
+            MessageLookupByLibrary.simpleMessage("Nghề nghiệp phù hợp"),
+        "ngon_ngu": MessageLookupByLibrary.simpleMessage("Ngôn ngữ"),
+        "noi_bat": MessageLookupByLibrary.simpleMessage("Nổi bật"),
+        "phat_trien_theo_so":
+            MessageLookupByLibrary.simpleMessage("Phát triển mạnh số"),
+        "so": MessageLookupByLibrary.simpleMessage("Con số"),
+        "than_so_hoc_la_gi":
+            MessageLookupByLibrary.simpleMessage("THẦN SỐ HỌC LÀ GÌ ?"),
+        "thiet_lap_ngay_sinh":
+            MessageLookupByLibrary.simpleMessage("Thiết lập ngày sinh"),
+        "tong_quan": MessageLookupByLibrary.simpleMessage("Tổng quan"),
+        "tu_vi_hom_nay": MessageLookupByLibrary.simpleMessage("Tử vi hôm nay"),
+        "tuoi": MessageLookupByLibrary.simpleMessage("Tuổi"),
+        "uu_diem": MessageLookupByLibrary.simpleMessage("Ưu điểm"),
+        "xem": MessageLookupByLibrary.simpleMessage("Xem"),
+        "xem_ngay_sinh": MessageLookupByLibrary.simpleMessage("Xem ngày sinh"),
+        "y_nghia_cac_con_so_tren_bieu_do": MessageLookupByLibrary.simpleMessage(
+            "MÝ NGHĨA CÁC CON SỐ TRÊN BIỂU ĐỒ")
+      };
 }
