@@ -7,7 +7,8 @@
 // ignore_for_file:unnecessary_brace_in_string_interps, unnecessary_new
 // ignore_for_file:prefer_single_quotes,comment_references, directives_ordering
 // ignore_for_file:annotate_overrides,prefer_generic_function_type_aliases
-// ignore_for_file:unused_import, file_names
+// ignore_for_file:unused_import, file_names, avoid_escaping_inner_quotes
+// ignore_for_file:unnecessary_string_interpolations, unnecessary_string_escapes
 
 import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
@@ -20,33 +21,40 @@ class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'zh';
 
   final messages = _notInlinedMessages(_notInlinedMessages);
-  static _notInlinedMessages(_) => <String, Function> {
-    "bieu_do_ngay_sinh" : MessageLookupByLibrary.simpleMessage("出生日期图"),
-    "chon_ngon_ngu" : MessageLookupByLibrary.simpleMessage("选择语言"),
-    "con_so_chu_dao_cua_ban_la" : MessageLookupByLibrary.simpleMessage("您的命运数字"),
-    "de_xuat" : MessageLookupByLibrary.simpleMessage("拟议的开发"),
-    "giai_thich" : MessageLookupByLibrary.simpleMessage("说明"),
-    "gioi_thieu" : MessageLookupByLibrary.simpleMessage("简介"),
-    "hay_chon_ngay_sinh_cua_ban" : MessageLookupByLibrary.simpleMessage("请选择您的出生日期"),
-    "home" : MessageLookupByLibrary.simpleMessage("家"),
-    "introl_1" : MessageLookupByLibrary.simpleMessage("命理学是根据您的姓名和出生日期来全面预测您的。通过命理学,您可以了解重要数字对您的生活有何影响。"),
-    "introl_2" : MessageLookupByLibrary.simpleMessage("该应用程序将分析您的姓名和生日,以眨眼即可显示结果！了解出生数字或生命线（表达您的性格和生活观）。让我们看看这些数字怎么说你"),
-    "khuyet_diem" : MessageLookupByLibrary.simpleMessage("弱点"),
-    "moc_cuoc_doi" : MessageLookupByLibrary.simpleMessage("生命的里程碑"),
-    "muc_dich_cuoc_song" : MessageLookupByLibrary.simpleMessage("生活目标"),
-    "nghe_nghiep_phu_hop" : MessageLookupByLibrary.simpleMessage("合适的职业"),
-    "ngon_ngu" : MessageLookupByLibrary.simpleMessage("语言"),
-    "noi_bat" : MessageLookupByLibrary.simpleMessage("精选"),
-    "phat_trien_theo_so" : MessageLookupByLibrary.simpleMessage("成长"),
-    "so" : MessageLookupByLibrary.simpleMessage("数字"),
-    "than_so_hoc_la_gi" : MessageLookupByLibrary.simpleMessage("什么是命理学？"),
-    "thiet_lap_ngay_sinh" : MessageLookupByLibrary.simpleMessage("set date of birth"),
-    "tong_quan" : MessageLookupByLibrary.simpleMessage("概述"),
-    "tu_vi_hom_nay" : MessageLookupByLibrary.simpleMessage("Horoscope today"),
-    "tuoi" : MessageLookupByLibrary.simpleMessage("岁岁"),
-    "uu_diem" : MessageLookupByLibrary.simpleMessage("优势"),
-    "xem" : MessageLookupByLibrary.simpleMessage("提交"),
-    "xem_ngay_sinh" : MessageLookupByLibrary.simpleMessage("首页"),
-    "y_nghia_cac_con_so_tren_bieu_do" : MessageLookupByLibrary.simpleMessage("说明图表的编号")
-  };
+  static Map<String, Function> _notInlinedMessages(_) => <String, Function>{
+        "bieu_do_ngay_sinh": MessageLookupByLibrary.simpleMessage("出生日期图"),
+        "chon_ngon_ngu": MessageLookupByLibrary.simpleMessage("选择语言"),
+        "con_so_chu_dao_cua_ban_la":
+            MessageLookupByLibrary.simpleMessage("您的命运数字"),
+        "de_xuat": MessageLookupByLibrary.simpleMessage("拟议的开发"),
+        "giai_thich": MessageLookupByLibrary.simpleMessage("说明"),
+        "gioi_thieu": MessageLookupByLibrary.simpleMessage("简介"),
+        "hay_chon_ngay_sinh_cua_ban":
+            MessageLookupByLibrary.simpleMessage("请选择您的出生日期"),
+        "home": MessageLookupByLibrary.simpleMessage("家"),
+        "introl_1": MessageLookupByLibrary.simpleMessage(
+            "命理学是根据您的姓名和出生日期来全面预测您的。通过命理学,您可以了解重要数字对您的生活有何影响。"),
+        "introl_2": MessageLookupByLibrary.simpleMessage(
+            "该应用程序将分析您的姓名和生日,以眨眼即可显示结果！了解出生数字或生命线（表达您的性格和生活观）。让我们看看这些数字怎么说你"),
+        "khuyet_diem": MessageLookupByLibrary.simpleMessage("弱点"),
+        "moc_cuoc_doi": MessageLookupByLibrary.simpleMessage("生命的里程碑"),
+        "muc_dich_cuoc_song": MessageLookupByLibrary.simpleMessage("生活目标"),
+        "nghe_nghiep_phu_hop": MessageLookupByLibrary.simpleMessage("合适的职业"),
+        "ngon_ngu": MessageLookupByLibrary.simpleMessage("语言"),
+        "noi_bat": MessageLookupByLibrary.simpleMessage("精选"),
+        "phat_trien_theo_so": MessageLookupByLibrary.simpleMessage("成长"),
+        "so": MessageLookupByLibrary.simpleMessage("数字"),
+        "than_so_hoc_la_gi": MessageLookupByLibrary.simpleMessage("什么是命理学？"),
+        "thiet_lap_ngay_sinh":
+            MessageLookupByLibrary.simpleMessage("set date of birth"),
+        "tong_quan": MessageLookupByLibrary.simpleMessage("概述"),
+        "tu_vi_hom_nay":
+            MessageLookupByLibrary.simpleMessage("Horoscope today"),
+        "tuoi": MessageLookupByLibrary.simpleMessage("岁岁"),
+        "uu_diem": MessageLookupByLibrary.simpleMessage("优势"),
+        "xem": MessageLookupByLibrary.simpleMessage("提交"),
+        "xem_ngay_sinh": MessageLookupByLibrary.simpleMessage("首页"),
+        "y_nghia_cac_con_so_tren_bieu_do":
+            MessageLookupByLibrary.simpleMessage("说明图表的编号")
+      };
 }
